Fix concurrent testcase random failure
Fix IUDConcurrentTest to run sql concurrently in correct order

This closes #1800
